Add ToNetworkOrder32 and ToNetworkOrder16
This patch adds ToNetworkOrder32 and ToNetworkOrder16 functions and
their tests to order.go and order_test.go respectively. ToNetworkOrder32
and ToNetworkOrder16 functions convert uint32 and uint16 values to the
ones in the network byte order. If the native endianess equals to the
network byte order, they return the original values.

Signed-off-by: Taku Fukushima <taku@soracom.jp>